+++ AVAFoundation/AFValidator.js
@@ -9,18 +9,21 @@ class AFValidator {
     this.invalid = { body: {}, query: {}, param: {} };
     request.body.validate = (key) => {
       this.key = key
+      this.label = key
       this.scope = "body";
       this.value = request.body[key];
       return this;
     }
     request.query.validate = (key) => {
       this.key = key
+      this.label = key
       this.scope = "query";
       this.value = request.query[key];
       return this;
     }
     request.params.validate = (key) => {
       this.key = key
+      this.label = key
       this.scope = "param";
       this.value = request.params[key];
       return this;
@@ -30,23 +33,32 @@ class AFValidator {
       for (const method of Object.keys(this.invalid)) {
         for (const key in this.invalid[method]) {
           for (const condition in this.invalid[method][key]) {
+            const label = this.invalid[method][key].label;
             const conditionValue = this.invalid[method][key][condition];
             switch (condition) {
-              case "type":
-                failures.push({ field: key, scope: method, error: "invalidType", message: `Type of '${key}' must be '${conditionValue}'.` });
+              case "vEquals":
+                failures.push({ field: key, scope: method, error: "unexpectedValue", message: `'${label}' is does not equal expected value.` });
                 break;
-              case "length":
-                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${key}' must be ${conditionValue}.` });
+              case "vNot":
+                failures.push({ field: key, scope: method, error: "isRestrictedValue", message: `'${label}' must not be restricted value.` });
                 break;
-              case "range":
-                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${key}' must be in range ${conditionValue}.` });
+              case "vNotEmpty":
+                failures.push({ field: key, scope: method, error: "isEmpty", message: `'${label}' must not be empty.` });
                 break;
-              case "greaterThan":
-                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${key}' must be greater than ${conditionValue}.` });
+              case "vType":
+                failures.push({ field: key, scope: method, error: "invalidType", message: `Type of '${label}' must be '${conditionValue}'.` });
                 break;
-              case "LessThan":
-                errors.push()
-                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${key}' must be less than ${conditionValue}.` });
+              case "vLength":
+                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${label}' must be ${conditionValue}.` });
+                break;
+              case "vRange":
+                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${label}' must be in range ${conditionValue}.` });
+                break;
+              case "vGreaterThan":
+                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${label}' must be greater than ${conditionValue}.` });
+                break;
+              case "vLessThan":
+                failures.push({ field: key, scope: method, error: "invalidLength", message: `Length of '${label}' must be less than ${conditionValue}.` });
                 break;
             }
           }
@@ -56,6 +68,41 @@ class AFValidator {
       onFailure({ error: new Error("Validation failed"), errors: failures });
       return false;
     }
+  }
+
+  alias(alias) {
+    this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
+    if (typeof alias === "string") {
+      this.invalid[this.scope][this.key].label = alias;
+    }
+    return this;
+  }
+
+  equals(value) {
+    if (value === this.value) {
+      return this;
+    }
+    this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
+    this.invalid[this.scope][this.key].vEquals = null;
+    return this;
+  }
+
+  not(value) {
+    if (value !== this.value) {
+      return this;
+    }
+    this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
+    this.invalid[this.scope][this.key].vNot = null;
+    return this;
+  }
+
+  notEmpty() {
+    if (typeof this.value !== "boolean" && this.value) {
+      return this;
+    }
+    this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
+    this.invalid[this.scope][this.key].vNotEmpty = null;
+    return this;
   }
 
   type(type) {
@@ -76,7 +123,7 @@ class AFValidator {
       return this;
     }
     this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
-    this.invalid[this.scope][this.key].type = type;
+    this.invalid[this.scope][this.key].vType = type;
     return this;
   }
 
@@ -93,24 +140,24 @@ class AFValidator {
       }
     }
     this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
-    this.invalid[this.scope][this.key].range = `${min}-${max}`;
+    this.invalid[this.scope][this.key].vRange = `${min}-${max}`;
     return this;
   }
 
   length(number) {
     if (Array.isArray(this.value) || typeof this.value === "string") {
       if (this.value.length == number) {
-        return true;
+        return this;
       }
     }
     if (typeof this.value === "number") {
       if (this.value == number) {
-        return true;
+        return this;
       }
     }
     this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
-    this.invalid[this.scope][this.key].length = number;
-    return false;
+    this.invalid[this.scope][this.key].vLength = number;
+    return this;
   }
 
   greaterThan(number) {
@@ -125,7 +172,7 @@ class AFValidator {
       }
     }
     this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
-    this.invalid[this.scope][this.key].greaterThan = number;
+    this.invalid[this.scope][this.key].vGreaterThan = number;
     return this;
   }
 
@@ -141,7 +188,7 @@ class AFValidator {
       }
     }
     this.invalid[this.scope][this.key] = Array.isArray(this.invalid[this.scope][this.key]) ? this.invalid[this.scope][this.key] : [];
-    this.invalid[this.scope][this.key].lessThan = number;
+    this.invalid[this.scope][this.key].vLessThan = number;
     return this;
   }